requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.42.0
django-appconf==1.0.4
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-imagekit==4.0.2
idna==2.10
oauthlib==3.1.0
pilkit==2.0
Pillow==8.0.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
